requirements.txt
python-slugify==7.0.0